ENH: move slic macros in separate header
